[수정] 내장 메서드로 FK 조회 가능여부 확인

+++ BackResale/src/test/java/com/golflearn/domain/repository/ResaleBoardRepositoryTest.java
@@ -1,26 +1,30 @@
 package com.golflearn.domain.repository;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import java.beans.Transient;
 import java.util.List;
 import java.util.Optional;
+
+import javax.transaction.Transactional;
 
 import org.junit.jupiter.api.Test;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.transaction.annotation.Transactional;
 
 import com.golflearn.domain.entity.ResaleBoardEntity;
+
 
 @SpringBootTest
 class ResaleBoardRepositoryTest {
 
 	@Autowired
 	private ResaleBoardRepository resaleBoardRepo;
+	
+	@Autowired
+	private ResaleCommentRepository resaleCmtRepo;
 	
 	Logger logger = LoggerFactory.getLogger(getClass());
 	
@@ -40,16 +44,17 @@ class ResaleBoardRepositoryTest {
 	// 상세 불러오기 (게시글 번호의 게시물을 조회)
 	@Test
 	void testFindDetail() {
-		Long resaleBoardNo = 26L;
-		ResaleBoardEntity resaleBoardEntity = resaleBoardRepo.findDetail(resaleBoardNo);
-		assertTrue(resaleBoardEntity != null);
+		Long resaleBoardNo = 1L;
+		resaleBoardRepo.findById(resaleBoardNo);
+		//		ResaleBoardEntity resaleBoardEntity = resaleBoardRepo.findDetail(resaleBoardNo);
+//		assertTrue(resaleBoardEntity != null);
 	}	
 	
 	// 조회수 증가
 	@Test
 	@Transient
 	void testViewCnt() {
-		Long resaleBoardNo = 1L;
+		Long resaleBoardNo = 3L;
 		Optional<ResaleBoardEntity> optRb = resaleBoardRepo.findById(resaleBoardNo); 
 		// isPresent() : boolean타입 
 		// ifPresent() : void 타입 . 값을 가지고 있는 지 확인 후 에러처리
@@ -93,11 +98,18 @@ class ResaleBoardRepositoryTest {
 	@Test
 //	@Transactional
 	void testDeleteBoard() {
-		Long resaleBoardNo = 43L;
-		resaleBoardRepo.findById(resaleBoardNo);
-		resaleBoardRepo.deleteComments(resaleBoardNo);
-		resaleBoardRepo.deleteLike(resaleBoardNo);
-		resaleBoardRepo.deleteById(resaleBoardNo);
+		Long resaleBoardNo = 76L;
+		Optional<ResaleBoardEntity> optRb = resaleBoardRepo.findById(resaleBoardNo);
+		if(optRb.isPresent()) {
+			resaleBoardRepo.deleteById(resaleBoardNo);
+			
+		}
+		
+//		resaleBoardRepo.deleteComments(resaleBoardNo);
+//		resaleBoardRepo.deleteLike(resaleBoardNo);
+//		resaleBoardRepo.deleteById(resaleBoardNo);
+//		int expectedCmtCnt = 0;
+//		assertEquals(expectedCmtCnt, optRb.get().getResaleBoardCmtCnt());
 	}
 	
 	// 게시글 검색
